[Mavproxy page]Step 1: Added all the mavproxy functions in a separate file and imported them. Status: Everything works!
